Make schema classes to allow for optional columns
